extract module benches::state_debug
